backend/requirements.txt
supabase
python-dotenv
Flask==3.0.0
flask-cors==4.0.0
torch
transformers
pillow
requests